[Editorial] Modernize hyphenation and spelling
There are several instances of the word "bi-sexual." The tool removed the hyphen, but I manually re-entered it. In these cases, the word does not have the modern meaning. When the text refers to a "bi-sexual race" etc., it means a race of people consisting of both men and women.