[Editorial] Changed misspellings present in original text
The original page scans as well as the Gutenberg text have the following misspellings, which I've chosen to change.

"Esparanto" in Ch. 3 (changed to "Esperanto")

"Alma" in Ch. 8 (changed to "Alima," as the character is referred to as Alima everywhere else in the text)

"Almwroth Wright" in Ch. 12 (changed to "Almroth Wright" as that is the name of the non-fictional person; his name is spelled this way elsewhere in the text as well)